Fix vendored builds through `[patch]`
The `[replace]` mechanism unfortunately didn't work for no-network style builds
because the implementation of the feature still required updating the git
repository found remotely. This commit updates the `Cargo.toml` of Rust to use
the `[patch]` feature in Cargo which should correctly handle this. The
`Cargo.lock` file now no longer mentions git repositories and should restore the
ability for no-network-style builds to work.

To ensure we don't regress this again the distcheck step has been updated to set
a dummy value for `CARGO_HOME` that for sure doesn't have anything in it, so if
it needs to get updated we'll generate an error due to `--frozen`.